requirements.txt
eth-brownie
eth-abi
eth-utils
black
numpy